Allow IP address detection to work without netroute
Because netroute is not available on Windows and I haven’t been able to
find another way to detect the default route there, see if we can find
a unique non-internal IPv4 address as a fallback. Also, improve error
reporting in case we aren’t able to detect the IP address.

I’ve removed the automatic retry after 2 seconds, because I felt it was
of limited use (if you lose the address you would also have lost the
default route?) and it would slow down error reporting in the common
case where the error persists.